% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boot-ns.r
\name{boot.ns}
\alias{boot.ns}
\title{Bootstrapping a Neymann-Scott point process model}
\usage{
boot.ns(fit, rchild, N, prog = TRUE)
}
\arguments{
\item{fit}{A fitted object from \link{fit.ns}().}

\item{rchild}{A function for the generation of random variables
from the distribution of the number of children for each
parent. The first argument of this function must be named \code{n},
and set the number of random values to generate. The second argument must be  Built in functions
(e.g., \code{rpois}) are suitable.}

\item{N}{The number of bootstrap resamples.}

\item{prog}{Logical, if \code{TRUE}, a progress bar is printed to
the console.}
}
\value{
The first argument, with added information from the
bootstrap procedure.
}
\description{
Carries out a bootstrap for Neymann-Scott point process models
fitted by \link{fit.ns}().
}

