% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ns.r
\name{boot}
\alias{boot}
\title{Bootstrapping for fitted models}
\usage{
boot(fit, N, prog = TRUE)
}
\arguments{
\item{fit}{A fitted object.}

\item{N}{The number of bootstrap resamples.}

\item{prog}{Logical, if \code{TRUE}, a progress bar is printed to
the console.}
}
\value{
The original model object containing additional information
    from the boostrap procedure. These are accessed by functions
    such as \link{summary}.

The original R6 reference object, with additional bootstrap
    information attached.
}
\description{
Carries out a parametric bootstrap procedure for models fitted
using this package.
}

