% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buffer_distances}
\alias{buffer_distances}
\title{Calculating distances between points subject to a buffer-zone edge correction.}
\usage{
buffer_distances(points, lims, R)
}
\arguments{
\item{points}{A matrix containing locations of observed points,
where each row corresponds to a point and each column
corresponds to a dimension.}

\item{lims}{A matrix with two columns, corresponding to the upper
and lower limits of each dimension, respectively.}

\item{R}{Truncation distance for the difference process.}
}
\description{
Calculates pairwise distances between points subject to a buffer-zone edge correction.
}

