% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.r
\name{coef.nspp}
\alias{coef.nspp}
\title{Extract Neyman-Scott point process parameter estiamtes}
\usage{
\method{coef}{nspp}(object, ...)
}
\arguments{
\item{object}{A fitted model from \link{fit.ns}().}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Extracts estimated and derived parameters from a model fitted using
\link{fit.ns}().
}

