% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{coef.nspp_r6}
\alias{coef.nspp_r6}
\title{Extract parameter estimates.}
\usage{
\method{coef}{nspp_r6}(object, ...)
}
\arguments{
\item{object}{A fitted model object returned by \link{fit.ns_r6}.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Extracts estimated parameters from an object returned by
\link{fit.ns_r6}.
}

