% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{coef.nspp_twoplanechild}
\alias{coef.nspp_twoplanechild}
\title{Extract parameter estimates.}
\usage{
\method{coef}{nspp_twoplanechild}(object, se = FALSE, report.2D = TRUE, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{se}{Logical, if \code{TRUE} standard errors are presented (if
available) instead of parameter estimates.}

\item{report.2D}{Logical, if \code{TRUE}, two-dimensional density
is reported.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Extracts estimated parameters from an object returned by
\link{fit.ns} with \code{"twoplane"} dispersion or
\link{fit.twoplane}.
}

