% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{coef.palm}
\alias{coef.palm}
\title{Extract parameter estimates.}
\usage{
\method{coef}{palm}(object, se = FALSE, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{se}{Logical, if \code{TRUE} standard errors are presented
(if available) instead of parameter estimates.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Extracts estimated parameters from an object returned by
\link{fit.ns} or \link{fit.void}.
}
