% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{coef.twoplane.nspp}
\alias{coef.twoplane.nspp}
\title{Extract two-plane survey parameter estimates}
\usage{
\method{coef}{twoplane.nspp}(object, all = FALSE, se = FALSE, ...)
}
\arguments{
\item{object}{A fitted model from \link{fit.twoplane}.}

\item{all}{Logical, if code{TRUE} derived parameters are also
included.}

\item{se}{Logical, if \code{TRUE} standard errors are presented (if
available) instead of parameter estimates.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Extracts estimated and derived parameters for a model fitted using
\link{fit.twoplane}.
}

