% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STLS-methods.R
\name{cv.fitSTLS}
\alias{cv.fitSTLS}
\title{cv.fitSTLS}
\usage{
cv.fitSTLS(X, y, nfolds = 5, foldid = NULL, center = TRUE,
  lambda = NULL, nlambda = 100, lmin_ratio = 1e-04, alpha = 1,
  eps_abs = 1e-04, eps_rel = 1e-04, maxit = 1000L, warm_start = TRUE)
}
\arguments{
\item{X}{The observation data matrix}

\item{y}{The observation response vector}

\item{nfolds}{Number of folds in cross-validation}

\item{foldid}{Folds id in N-fold cross-validation}

\item{center}{Y = Y - mean(Y) and X = X - mean(X)}

\item{lambda}{The regularized lambda vector}

\item{nlambda}{The Number of lambda}

\item{lmin_ratio}{lambda_max / lambda_min}

\item{alpha}{alpha parameter for elastic net penalty (0,1](Ridge Regression --> Lasso)}

\item{eps_abs}{Absolutely epsilon for generated varaince vector}

\item{eps_rel}{Relative epsilon for objective function value of generated vector}

\item{maxit}{The maximal iteration of PALM algorithm}

\item{warm_start}{Warm start for regularizer parameter tuning}
}
\value{
\code{STLS_fit} object
}
\description{
Cross-validation for Sparse Total Least Square model Using PALM algorithm
}

