% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{empirical.palm}
\alias{empirical.palm}
\title{Plotting the empirical Palm intensity}
\usage{
empirical.palm(points, lims, breaks = 50, xlim = NULL, add = FALSE, ...)
}
\arguments{
\item{points}{A matrix containing locations of observed points,
where each row corresponds to a point and each column corresponds
to a dimension.}

\item{lims}{A matrix with two columns, corresponding to the upper
and lower limits of each dimension, respectively.}

\item{breaks}{The (approximate) number of points plotted.}

\item{xlim}{The x-axis limits for the plot.}

\item{add}{Logical, if \code{TRUE} then the line is added to a
plot.}

\item{...}{Graphical parameters (e.g., to be passed to
\link{par}().}
}
\value{
A plot showing the empirical Palm intensity.
}
\description{
Plots the Palm intensity from empirical data.
}
\examples{
empirical.palm(example.2D, lims = rbind(c(0, 1), c(0, 1)), breaks = 25)

}

