% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit-ns.r
\docType{data}
\name{example.1D}
\alias{example.1D}
\title{1-dimensional example data}
\format{A matrix.}
\usage{
example.1D
}
\description{
Simulated data, with children points generated from a Binomial(4,
0.5) distribution.
}
\keyword{datasets}

