% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ns.r
\docType{data}
\name{example.2D}
\alias{example.2D}
\title{2-dimensional example data}
\format{A matrix.}
\usage{
example.2D
}
\description{
Simulated data, with children points generated from a Binomial(2,
0.5) distribution.
}
\keyword{datasets}

