% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm.r
\docType{data}
\name{example.twocamera}
\alias{example.twocamera}
\title{Two-camera example data.}
\format{
A list.
}
\usage{
example.twocamera
}
\description{
Simulated data from a two-camera aerial survey.
}
\keyword{datasets}
