% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ns.r
\docType{data}
\name{example.twoplane}
\alias{example.twoplane}
\title{Two-plane example data.}
\format{A list.}
\usage{
example.twoplane
}
\description{
Simulated data from a two-plane aerial survey.
}
\keyword{datasets}

