% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ns.r
\name{fit.ns_r6}
\alias{fit.ns_r6}
\title{Fitting a model to a Neyman-Scott point process}
\usage{
fit.ns_r6(points, lims, R, disp = "gaussian", child.dist = "pois",
  edge.correction = "pbc", siblings = NULL, trace = FALSE)
}
\arguments{
\item{points}{A matrix containing locations of observed points,
where each row corresponds to a point and each column
corresponds to a dimension.}

\item{lims}{A matrix with two columns, corresponding to the upper
and lower limits of each dimension, respectively.}

\item{R}{Truncation distance for the difference process.}

\item{disp}{A character string, providing the distribution for the
dispersion of points around the parents. This can either be
\code{"gaussian"} for Gaussian dispersion, or \code{"matern"}
for uniform dispersion (giving a matern process).}

\item{child.dist}{The distribution of the number of children
generated by a randomly selected parent; for a Poisson
distribution, use \code{"pois"}.}

\item{edge.correction}{The method used for the correction of edge
effects. Either \code{"pbc"} for periodic boundary conditions,
or \code{"buffer"} for a buffer-zone correction.}

\item{siblings}{A named list, containing the following three
components: (1) A component named \code{matrix}, where the jth
element of the ith row is \code{TRUE} if the ith and jth
observed points are known siblings, \code{FALSE} if they are
known non-siblings, and \code{NA} if it is not known whether or
not they are siblings, (2) A component named pT, containing a
scalar providing the probability that the element (i, j) in the
component \code{matrix} is \code{TRUE}, conditional on the ith
and jth points being siblings, (3) A component named pF,
containing a scalar providing the probability that the element
(i, j) in the component \code{matrix} is \code{FALSE},
conditional on the ith and jth points not being
siblings. Defaults to an argument representing a Poisson
distribution.}

\item{trace}{Logical, if \code{TRUE}, parameter values are printed
to the screen for each iteration of the optimisation procedure.}
}
\value{
An R6 reference class object.
}
\description{
Estimates parameters for a Neyman-Scott point process using the
approach of Tanaka (2008).
}
\references{
Tanaka, U., Ogata, Y., and Stoyan, D. (2008) Parameter
estimation and model selection for Neyman-Scott point
processes. \emph{Biometrical Journal}, \strong{50}: 43--57.
}

