% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ns.r
\name{fit.twoplane_r6}
\alias{fit.twoplane_r6}
\title{Estimation of animal density from two-plane surveys.}
\usage{
fit.twoplane_r6(points, planes, d, w, b, l, tau, R, edge.correction = "pbc",
  start = NULL, bounds = NULL, trace = FALSE)
}
\arguments{
\item{points}{A vector (or single-column matrix) containing the
distance along the transect that each detection was made.}

\item{planes}{A vector containing the plane ID (either \code{1} or
\code{2}) that made the corresponding detection in
\code{points}.}

\item{d}{The length of the transect flown (in km).}

\item{w}{The distance from the transect to which detection of
individuals on the surface is certain. This is equivalent to
the half-width of the detection zone.}

\item{b}{The distance from the transect to the edge of the area of
interest. Conceptually, the distance between the transect and
the furthest distance a whale could be on the passing on the
first plane and plausibly move into the detection zone by the
passing of the second plane.}

\item{l}{The lag between planes (in seconds).}

\item{tau}{Mean dive-cycle duration (in seconds).}

\item{R}{Truncation distance (see \link{fit.ns}).}

\item{edge.correction}{The method used for the correction of edge
effects. Either \code{"pbc"} for periodic boundary conditions,
or \code{"buffer"} for a buffer-zone correction.}

\item{start}{A named vector of starting values for the model
parameters.}

\item{bounds}{A list with named components. Each component should
be a vector of length two, giving the upper and lower bounds
for the named parameter.}

\item{trace}{Logical, if \code{TRUE}, parameter values are printed
to the screen for each iteration of the optimisation procedure.}
}
\value{
An object of class \code{"nspp_r6"} that can be extracted via
the same utility functions fit for objects created using
\code{fit.ns_6()}.
}
\description{
Estimates animal density (amongst other parameters) from two-plane
aerial surveys. This conceptualises sighting locations as a
Neyman-Scott point pattern---estimation is carried out via
\code{fit.ns_r6()}.
}

