% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PALM-package.R
\docType{package}
\name{PALM}
\alias{PALM}
\alias{PALM-package}
\alias{package-PALM}
\title{Solving noncovex and nonsmooth optimization problems using the
Proximal Alternating Linearized Minization (PALM) algorithm}
\description{
Solving noncovex and nonsmooth optimization problems using the
Proximal Alternating Linearized Minization (PALM) algorithm
}

