% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STLS-methods.R
\name{plot.cv.STLS_fit}
\alias{plot.cv.STLS_fit}
\title{plot cross-validation curive and search for lambda.min}
\usage{
\method{plot}{cv.STLS_fit}(x, ...)
}
\arguments{
\item{x}{fitted cv.STLS_fit}
}
\description{
plot cross-validation curive and search for lambda.min
}

