% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot.nspp}
\alias{plot.nspp}
\title{Plotting an estimated Palm intensity function.}
\usage{
\method{plot}{nspp}(x, plot.empirical = FALSE, breaks = NULL, xlim = NULL,
  ylim = NULL, ...)
}
\arguments{
\item{x}{A fitted model from \link{fit.ns}().}

\item{plot.empirical}{Logical, if \code{TRUE} then the empirical
Palm intensity is also plotted.}

\item{breaks}{The (approximate) number of points plotted.}

\item{xlim}{The x-axis limits for the plot.}

\item{ylim}{The y-axis limits for the plot.}

\item{...}{Graphical parameters (e.g., to be passed to
\link{par}().}
}
\description{
Plots a fitted Palm intensity function from an object returned by
\link{fit.ns}().
}
\examples{
## Fitting a model.
fit <- fit.ns(example.2D, lims = rbind(c(0, 1), c(0, 1)), R = 0.5)
## Plotting.
plot(fit, plot.empirical = TRUE, breaks = 30, ylim = c(0, 200))

}

