% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot.nspp_r6}
\alias{plot.nspp_r6}
\title{Plotting an estimated Palm intensity function.}
\usage{
\method{plot}{nspp_r6}(x, ...)
}
\arguments{
\item{x}{A fitted model from \link{fit.ns_r6}.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Plots a fitted Palm intensity function from an object returned by
\link{fit.ns_r6}().
}

