% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{plot.palm}
\alias{plot.palm}
\title{Plotting an estimated Palm intensity function.}
\usage{
\method{plot}{palm}(x, ...)
}
\arguments{
\item{x}{A fitted model from \link{fit.ns}.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Plots a fitted Palm intensity function from an object returned by
\link{fit.ns}().
}

