% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{poly_to_raster}
\alias{poly_to_raster}
\title{Convert polygons to raster}
\usage{
poly_to_raster(p, z, rows = 100, cols = 100)
}
\description{
Takes an object of class "SpatialPolygons" and converts to raster image. The argument \code{z} gives the value of each polygon that will be used when generating the raster.
}
