% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{poly_to_segment}
\alias{poly_to_segment}
\title{Convert polygons to segments}
\usage{
poly_to_segment(p)
}
\description{
Takes an object of class "SpatialPolygons" and extracts the coordinates of all polygons. Output is in the form of a matrix with columns "x0", "y0", "x1", "y1", corresponding to the start and end coordinates of each edge. These can be plotted using the \code{segments()} function, which is often faster than plotting raw polygons.
}
