% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling-matrix.r
\name{siblings.twoplane}
\alias{siblings.twoplane}
\title{Creating a sibling matrix for two-plane surveys}
\usage{
siblings.twoplane(plane.id)
}
\arguments{
\item{plane.id}{A vector indicating which plane collected each
observed location.}
}
\value{
A list, a suitable object for the \code{siblings} argument
to \link{fit.ns}.
}
\description{
Creates an object suitable for the \code{siblings} argument to
\link{fit.ns} for two-plane surveys of whale populations. There is
partial sibling information in that observations made by the same
plane cannot be siblings.
}

