% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim-ns.r
\name{sim.ns}
\alias{sim.ns}
\title{Simulating Neyman-Scott point process data}
\usage{
sim.ns(pars = NULL, lims = rbind(c(0, 1), c(0, 1)), rchild = rpois,
  non.siblings = NULL, plot.points = FALSE, plot.empirical = FALSE)
}
\arguments{
\item{pars}{A named vector (or list?) of parameter values. Required
parameters are \code{D}, the density of parent points;
\code{sigma}, controlling the dispersion of children around their
parents; and \code{child.par}, the parameter characterising the distribution for the number of children generated by each parent.}

\item{lims}{A matrix with two colums, corresponding to the upper
and lower limits of each dimension, respectively.}

\item{rchild}{A function for the generation of random variables
from the distribution of the number of children for each
parent. The first argument of this function must be named \code{n},
and set the number of random values to generate. The second argument must be  Built in functions
(e.g., \code{rpois}) are suitable.}

\item{non.siblings}{An argument specifying the type of sibling
information to return (not sure what this should look like yet).}

\item{plot.points}{Logical, if \code{TRUE}, simulated parent and children
point locations will be plotted (only for two dimensions).}

\item{plot.empirical}{Logical, if \code{TRUE}, the empirical Palm
intensity is plotted, with the true Palm intensity from the
provided parameters overlain. The latter is only approximate as the
mean and variance of the number of children per parent are
calculated via simulation from rchild.}

\item{...}{Further parameters for rchild.dist.}
}
\value{
A matrix containing simulated point locations.
}
\description{
Simulates point locations from a Neyman-Scott point process.
}

