% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-ns.r
\name{sim.twoplane.2D}
\alias{sim.twoplane.2D}
\title{Simulating two-dimensional two-plane whale survey data}
\usage{
sim.twoplane.2D(pars, lims, w, b)
}
\arguments{
\item{pars}{A named vector of parameter values. Required parameters
are \code{D}, whale density, \code{sigma}, whale movement,
\code{p.up.down}, the probability that a whale is on the
surface when the second plane flies over, given that it was
submerged when the first plane flew over, and \code{p.down.up},
the probability that a whale is submerged with the second plane
flies over, given that it was on the surface when the first
plane flew over.}

\item{lims}{The one-dimensional limits of the transect.}

\item{w}{The distance from the transect and the edge of the
detection zone.}

\item{b}{The distance from the transect and the edge of the buffer
zone.}
}
\value{
A list containing observed whale locations and associated
    plane IDs.
}
\description{
Simulates observed whale locations and plane IDs from a two-plane
whale survey.
}

