% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim-ns.r
\name{sim.twoplane}
\alias{sim.twoplane}
\title{Simulating two-plane whale survey data}
\usage{
sim.twoplane(pars, lims)
}
\arguments{
\item{pars}{A named vector of parameter values. Required parameters
are \code{D}, whale density, \code{sigma}, whale movement,
\code{p01}, the probability that a whale is on the surface when the
second plane flies over, given that it was submerged when the first
plane flew over, and \code{10}, the probability that a whale is
submerged with the second plane flies over, given that it was on
the surface when the first plane flew over.}

\item{lims}{The limits of the survey transect.}
}
\value{
A list containing observed whale locations and associated
plane IDs.
}
\description{
Simulates observed whale locations and plane IDs from a two-plane
whale survey.
}

