% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STLS-methods.R
\name{ss.fitSTLS}
\alias{ss.fitSTLS}
\title{ss.fitSTLS}
\usage{
ss.fitSTLS(X, y, nperm = 100, center = TRUE, lambda = NULL,
  nlambda = 100, lmin_ratio = 1e-04, alpha = 1, eps_abs = 1e-04,
  eps_rel = 1e-04, maxit = 1000L, warm_start = TRUE)
}
\arguments{
\item{X}{The observation data matrix}

\item{y}{The observation response vector}

\item{nperm}{Number of permutaiton of stabilioty-selection}

\item{center}{Y = Y - mean(Y) and X = X - mean(X)}

\item{lambda}{The regularized lambda vector}

\item{nlambda}{The Number of lambda}

\item{lmin_ratio}{lambda_max / lambda_min}

\item{alpha}{alpha parameter for elastic net penalty (0,1](Ridge Regression --> Lasso)}

\item{eps_abs}{Absolutely epsilon for generated varaince vector}

\item{eps_rel}{Relative epsilon for objective function value of generated vector}

\item{maxit}{The maximal iteration of PALM algorithm}

\item{warm_start}{Warm start for regularizer parameter tuning}
}
\value{
\code{STLS_fit} object
}
\description{
Stability-Selection for Sparse Total Least Square model Using PALM algorithm
}

