% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{summary.palm}
\alias{summary.palm}
\title{Summarising palm model fits}
\usage{
\method{summary}{palm}(object, ...)
}
\arguments{
\item{object}{A fitted model returned by \link{fit.ns}.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Provides a useful summary of the model fit.
}
