% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sibling-matrix.r
\name{twoplane.siblings}
\alias{twoplane.siblings}
\title{Creating a sibling matrix for two-plane surveys}
\usage{
twoplane.siblings(plane.id)
}
\arguments{
\item{plane.id}{A vector indicating which plane collected each
observed location.}
}
\value{
A list, a suitable object for the \code{siglings} argument
to \link{ns.fit}.
}
\description{
Creates an object suitable for the \code{siblings} argument to
\link{fit.ns} for two-plane surveys of whale populations. There is
partial sibling information in that observations made by the same
plane cannot be siblings.
}

