% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boot-ns.r
\name{boot.ns}
\alias{boot.ns}
\title{Bootstrapping a Neymann-Scott point process model}
\usage{
boot.ns(fit, rchild.fun, N, prog = TRUE)
}
\arguments{
\item{fit}{A fitted object from \link{fit.ns}().}

\item{rchild.fun}{A function for random number generation of the
number of children spawned by each parent. It must take two
arguments: The first argument must be \code{n}, the number of
random values to draw, and the second must be \code{child.par}, the
parameter specified to be estimated in the argument
\code{child.dist} argument to \link{fit.ns}(). The mean and
variance functions specified in this argument should be the
analytic mean and variance of the random numbers generated by
\code{rchild.fun}().}

\item{N}{The number of bootstrap resamples.}

\item{prog}{Logical, if \code{TRUE}, a progress bar is printed to
the console.}
}
\value{
The first argument, with added information from the
bootstrap procedure.
}
\description{
Carries out a bootstrap for Neymann-Scott point process models
fitted by \link{fit.ns}().
}

