% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-ns.r
\name{boot.ns}
\alias{boot.ns}
\title{Bootstrapping a Neymann-Scott point process model}
\usage{
boot.ns(fit, rchild, N, prog = TRUE)
}
\arguments{
\item{fit}{A fitted object from \link{fit.ns}().}

\item{rchild}{A function for the generation of random variables
from the distribution of the number of children for each
parent. The first argument of this function must set the number of
random values to generate. The second argument must be the
parameter characterising the distribution for the number of
children generated by each parent. See 'Details' for more
information.}

\item{N}{The number of bootstrap resamples.}

\item{prog}{Logical, if \code{TRUE}, a progress bar is printed to
the console.}
}
\value{
The first argument, with added information from the
bootstrap procedure.
}
\description{
Carries out a bootstrap for Neymann-Scott point process models
fitted by \link{fit.ns}.
}
\details{
The \code{rchild} function may only take a single distributional
parameter. If the distribution for the number of children generated
by each parent is Poisson, then the native \code{rpois} is
appropriate, as this distribution has a single parameter. For
distributions with two or more parameters, those other than
\code{child.par} must be hard-coded into \code{rchild}. For
example, if a Binomial(n, 2, p) is required, then \code{function(n,
p) rbinom(n = n, size = 2, prob = p)} would be an appropriate
function for \code{rchild}.
}
\examples{
\dontrun{
## Fitting a model.
fit <- fit.ns(example.2D, lims = rbind(c(0, 1), c(0, 1)), R = 0.5)
## Bootstrapping.
fit.boot <- boot.ns(fit, rchild = rpois, N = 100)
}

}

