% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{coef.nspp}
\alias{coef.nspp}
\title{Extract Neyman-Scott point process parameter estimates}
\usage{
\method{coef}{nspp}(object, all = FALSE, ...)
}
\arguments{
\item{object}{A fitted model from \link{fit.ns}.}

\item{all}{Logical, if \code{TRUE} derived parameters are also
included.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Extracts estimated and derived parameters from a model fitted using
\link{fit.ns}.
}

