% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{coef.palm}
\alias{coef.palm}
\alias{coef.palm_twocamerachild}
\title{Extract parameter estimates.}
\usage{
\method{coef}{palm}(object, se = FALSE, ...)

\method{coef}{palm_twocamerachild}(object, se = FALSE, report.2D = TRUE, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{se}{Logical, if \code{TRUE} standard errors are presented
(if available) instead of parameter estimates.}

\item{...}{Other parameters (for S3 generic compatibility).}

\item{report.2D}{Logical, for two-camera model fits only. If
\code{TRUE}, two-dimensional animal density is reported.}
}
\description{
Extracts estimated parameters from an object returned by the
fitting functions in this package, such as \link{fit.ns},
\link{fit.void}, and \link{fit.twocamera}.
}
