% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.r
\name{empirical.palm}
\alias{empirical.palm}
\title{Plotting the empirical Palm intensity}
\usage{
empirical.palm(points, lims, breaks = NULL, add = FALSE)
}
\arguments{
\item{points}{A matrix containing locations of observed points,
where each row corresponds to a point and each column corresponds
to a dimension.}

\item{lims}{A matrix with two columns, corresponding to the upper
and lower limits of each dimension, respectively.}

\item{breaks}{The (approximate) number of points plotted.}

\item{add}{Logical, if \code{TRUE} then the line is added to a
plot.}
}
\value{
A plot showing the empirical Palm intensity.
}
\description{
Plots the Palm intensity from empirical data.
}

