% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm.r
\docType{data}
\name{example.1D}
\alias{example.1D}
\title{1-dimensional example data}
\format{
A matrix.
}
\usage{
example.1D
}
\description{
Simulated data from a Neyman-Scott point process, with children
points generated in the interval [0, 1]. The number of children
spawned by each parent is from a Binomial(4, 0.5) distribution.
}
\keyword{datasets}
