% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit-ns.r
\name{fit.ns}
\alias{fit.ns}
\title{Fitting a model to a Neyman-Scott point process}
\usage{
fit.ns(points = NULL, lims = NULL, R, sigma.sv = 0.1 * R,
  sigma.bounds = c(0, R), child.dist = list(mean = function(x) x, var =
  function(x) x, sv = 5, bounds = c(1e-08, 1e+08)), non.siblings = NULL)
}
\arguments{
\item{points}{A matrix containing locations of observed points,
where each row corresponds to a point and each column corresponds
to a dimension.}

\item{lims}{A matrix with two columns, corresponding to the upper
and lower limits of each dimension, respectively.}

\item{R}{Truncation distance for the difference process.}

\item{sigma.sv}{The start value for \code{sigma} in optimisation.}

\item{sigma.bounds}{The bounds of the \code{sigma} parameter in
optimisation.}

\item{child.dist}{An argument describing the distribution
generating the number of children per parent. It must be a list
with four components: (1) A component named \code{mean} that
provides a function returning the expectation of the distribution,
as an argument of a single parameter to be estimated, (2) A
component named \code{var} that provides a function returning the
variance of a distribution, as an argument of a singla parameter to
be estimated, (3) A component named \code{sv} providing the start
value for the parameter to be estimated, and (4) A component named
\code{bounds} providing a vector of length two that gives the
parameter bounds.}

\item{non.siblings}{An object containing information about known
siblings (or known non-siblings). Not yet implemented.}

\item{sv}{A vector (or list?) of start values for optimisation.}
}
\value{
An object with information that can be extracted via other
utility functions.
}
\description{
Estimates parameters for a Neyman-Scott point process using the
approach of Tanaka (2008).
}
\references{
Tanaka, U., Ogata, Y., and Stoyan, D. (2008) Parameter
estimation and model selection for Neyman-Scott point
processes. \emph{Biometrical Journal}, \strong{50}: 43--57.
}

