% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit-ns.r
\name{fit.ns}
\alias{fit.ns}
\title{Fitting a model to a Neyman-Scott point process}
\usage{
fit.ns(points = NULL, lims = NULL, child.dist = NULL,
  non.siblings = NULL, sv = NULL, fix = NULL)
}
\arguments{
\item{points}{A matrix containing locations of observed points,
where each row corresponds to a point and each column corresponds
to a dimension.}

\item{lims}{A matrix with two colums, corresponding to the upper
and lower limits of each dimension, respectively.}

\item{child.dist}{A list containing information about the
distribution of the number of children generated by the parents. It
must contain components containing the following, in the same order
listed: (1) A vector of parameter names, (2) A function with
arguments corresponding to these parameter names, returning the
expectation, and (3) A function with arguments corresponding to
these parameter names, returning the variance. Note that only two
parameters can be estimated for this distribution.}

\item{non.siblings}{An object containing information about known
siblings (or known non-siblings).}

\item{sv}{A vector (or list?) of start values for optimisation.}

\item{fix}{A vector (or list?) of parameter values to fix.}
}
\value{
An object with information that can be extracted via other
utility functions.
}
\description{
Estimates parameters for a Neyman-Scott point process using the
approach of Tanaka (2008).
}
\references{
Tanaka, U., Ogata, Y., and Stoyan, D. (2008) Parameter
estimation and model selection for Neyman-Scott point
processes. \emph{Biometrical Journal}, \strong{50}: 43--57.
}

