% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ns.r
\name{fit.ns}
\alias{fit.ns}
\title{Fitting a model to a Neyman-Scott point process}
\usage{
fit.ns(points = NULL, lims = NULL, R, sigma.sv = 0.1 * R,
  sigma.bounds = c(1e-10, R), child.dist = list(mean = function(x) x, var =
  function(x) x, sv = 5, bounds = c(1e-08, 1e+08)), edge.correction = "pbc",
  siblings = NULL, trace = FALSE)
}
\arguments{
\item{points}{A matrix containing locations of observed points,
where each row corresponds to a point and each column
corresponds to a dimension.}

\item{lims}{A matrix with two columns, corresponding to the upper
and lower limits of each dimension, respectively.}

\item{R}{Truncation distance for the difference process.}

\item{sigma.sv}{The start value for \code{sigma} in optimisation.}

\item{sigma.bounds}{The bounds of the \code{sigma} parameter in
optimisation.}

\item{child.dist}{An argument describing the distribution
generating the number of children per parent. It must be a list
with four components: (1) A component named \code{mean} that
provides a function returning the expectation of the
distribution, with its only argument being the single parameter
to be estimated, (2) A component named \code{var} that provides
a function returning the variance of a distribution, with its
only argument being the single parameter to be estimated, (3) A
component named \code{sv} providing the start value for the
parameter to be estimated, and (4) A component named
\code{bounds} providing a vector of length two that gives the
parameter bounds.}

\item{edge.correction}{The method used for the correction of edge
effects. Either \code{"pbc"} for periodic boundary conditions,
or \code{"buffer"} for a buffer-zone correction.}

\item{siblings}{A named list, containing the following three
components: (1) A component named \code{matrix}, where the jth
element of the ith row is \code{TRUE} if the ith and jth
observed points are known siblings, \code{FALSE} if they are
known non-siblings, and \code{NA} if it is not known whether or
not they are siblings, (2) A component named pT, containing a
scalar providing the probability that the element (i, j) in the
component \code{matrix} is \code{TRUE}, conditional on the ith
and jth points being siblings, (3) A component named pF,
containing a scalar providing the probability that the element
(i, j) in the component \code{matrix} is \code{FALSE},
conditional on the ith and jth points not being
siblings. Defaults to an argument representing a Poisson
distribution.}

\item{trace}{Logical, if \code{TRUE}, parameter values are printed
to the screen for each iteration of the optimisation procedure.}
}
\value{
An object with information that can be extracted via other
utility functions.
}
\description{
Estimates parameters for a Neyman-Scott point process using the
approach of Tanaka (2008).
}
\examples{
## Poisson number of children per parent.
fit.2D.pois <- fit.ns(example.2D, lims = rbind(c(0, 1), c(0, 1)), R = 0.5)
## Binomial number of children per parent.
fit.1D.binom <- fit.ns(example.1D, lims = cbind(0, 1), R = 0.5,
                      child.dist = list(mean = function(x) 4*x,
                          var = function(x) 4*x*(1 - x),
                          sv = 0.5, bounds = c(0, 1)))

}
\references{
Tanaka, U., Ogata, Y., and Stoyan, D. (2008) Parameter
estimation and model selection for Neyman-Scott point
processes. \emph{Biometrical Journal}, \strong{50}: 43--57.
}

