% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ns.r
\name{fit.ns_r6}
\alias{fit.ns_r6}
\title{Fitting a model to a Neyman-Scott point process}
\usage{
fit.ns_r6(points, lims, R, disp = "gaussian", child.dist = "pois",
  child.info = NULL, sibling.list = NULL, edge.correction = "pbc",
  start = NULL, bounds = NULL, trace = FALSE)
}
\arguments{
\item{points}{A matrix containing locations of observed points,
where each row corresponds to a point and each column
corresponds to a dimension.}

\item{lims}{A matrix with two columns, corresponding to the upper
and lower limits of each dimension, respectively.}

\item{R}{Truncation distance for the difference process.}

\item{disp}{A character string indicating the distribution of
children around their parents. Use \code{"gaussian"} for
multivariate normal dispersion, or \code{"uniform"} for
uniform dispersion within distance \code{tau} of the parent.}

\item{child.dist}{The distribution of the number of children
generated by a randomly selected parent. For a Poisson
distribution, use \code{"pois"}; for a binomial distribution,
use \code{"binomx"}, where \code{"x"} is replaced by the fixed
value of the number of independent trials (e.g.,
\code{"binom5"} for a Binomial(5, p) distribution, and
\code{"binom50"} for a Binomial(50, p) distribution); and
\code{"twoplane"} for a child distribution appropriate for a
two-plane aerial survey.}

\item{child.info}{A list of further information that is required
about the distribution for the number of children generated by
parents. See `Details'.}

\item{sibling.list}{An optional list that comprises (i) a component
named \code{sibling.mat}, containing a matrix such that the jth
entry in the ith row is \code{TRUE} if the ith and jth points
are known siblings, \code{FALSE} if they are known nonsiblings,
and \code{NA} if their sibling status is not known; (ii) alpha,
providing the probability that a sibling is successfully
identified as a sibling; and (iii) beta, providing the
probability that a nonsibling is successfully identified as a
nonsibling.}

\item{edge.correction}{The method used for the correction of edge
effects. Either \code{"pbc"} for periodic boundary conditions,
or \code{"buffer"} for a buffer-zone correction.}

\item{start}{A named vector of starting values for the model
parameters.}

\item{bounds}{A list with named components. Each component should
be a vector of length two, giving the upper and lower bounds
for the named parameter.}

\item{trace}{Logical; if \code{TRUE}, parameter values are printed
to the screen for each iteration of the optimisation procedure.}
}
\value{
An R6 reference class object.
}
\description{
Estimates parameters for a Neyman-Scott point process using the
approach of Tanaka (2008).
}
\details{
The \code{"child.info"} argument is required when \code{child.dist}
is set to \code{"twoplane"}. It must be a list that comprises (i) a
component named \code{w}, providing the halfwidth of the detection
zone; (ii) a component named \code{b}, providing the halfwidth of
the survey area; (iii) a component named \code{l}, providing the
time lag between planes (in seconds); and (iv) a component named
\code{tau}, providing the mean dive-cycle duration.
}
\references{
Tanaka, U., Ogata, Y., and Stoyan, D. (2008) Parameter
estimation and model selection for Neyman-Scott point
processes. \emph{Biometrical Journal}, \strong{50}: 43--57.
}

