% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ns.r
\name{fit.twocamera}
\alias{fit.twocamera}
\title{Estimation of animal density from two-camera surveys.}
\usage{
fit.twocamera(
  points,
  cameras = NULL,
  d,
  w,
  b,
  l,
  tau,
  R,
  edge.correction = "pbc",
  start = NULL,
  bounds = NULL,
  trace = FALSE
)
}
\arguments{
\item{points}{A vector (or single-column matrix) containing the
distance along the transect that each detection was made.}

\item{cameras}{An optional vector containing the camera ID (either
\code{1} or \code{2}) that made the corresponding detection in
\code{points}.}

\item{d}{The length of the transect flown (in km).}

\item{w}{The distance from the transect to which detection of
individuals on the surface is certain. This is equivalent to
the half-width of the detection zone.}

\item{b}{The distance from the transect to the edge of the area of
interest. Conceptually, the distance between the transect and
the furthest distance a whale could be on the passing of the
first camera and plausibly move into the detection zone by the
passing of the second camera.}

\item{l}{The lag between cameras (in seconds).}

\item{tau}{Mean dive-cycle duration (in seconds).}

\item{R}{Truncation distance (see \link{fit.ns}).}

\item{edge.correction}{The method used for the correction of edge
effects. Either \code{"pbc"} for periodic boundary conditions,
or \code{"buffer"} for a buffer-zone correction.}

\item{start}{A named vector of starting values for the model
parameters.}

\item{bounds}{A list with named components. Each component should
be a vector of length two, giving the upper and lower bounds
for the named parameter.}

\item{trace}{Logical; if \code{TRUE}, parameter values are printed
to the screen for each iteration of the optimisation procedure.}
}
\value{
An R6 reference class object.
}
\description{
Estimates animal density (amongst other parameters) from two-camera
aerial surveys. This conceptualises sighting locations as a
Neyman-Scott point pattern.
}
\details{
This function is simply a wrapper for \code{fit.ns}, and
facilitates the fitting of the model proposed by Stevenson,
Borchers, and Fewster (2019). This function presents the
parameter \code{D.2D} (two-dimensional cetacean density in
cetaceans per square km) rather than \code{D} for enhanced
interpretability.

For further details on the cluster capture-recapture estimation
approach, see Fewster, Stevenson and Borchers (2016).
}
\examples{
## Fitting model.
fit <- fit.twocamera(points = example.twocamera$points, cameras = example.twocamera$cameras,
                     d = 500, w = 0.175, b = 0.5, l = 20, tau = 110, R = 1)
## Printing estimates.
coef(fit)
## Plotting the estimated Palm intensity.
plot(fit)

}
\references{
Fewster, R. M., Stevenson, B. C., and Borchers,
    D. L. (2016) Trace-contrast methods for capture-recapture
    without capture histories. \emph{Statistical Science},
    \strong{31}: 245--258.

Stevenson, B. C., Borchers, D. L., and Fewster,
    R. M. (2019) Cluster capture-recapture to account for
    identification uncertainty on aerial surveys of animal
    populations. \emph{Biometrics}.
}
\seealso{
Use \link{coef.palm} to extract estimated parameters, and
    \link{plot.palm} to plot the estimated Palm intensity
    function. Use \link{boot.palm} to run a parametric bootstrap,
    allowing calculation of standard errors and confidence
    intervals.

See \link{sim.twocamera} to simulate sightings from a
    two-camera aerial survey.
}
