% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pbc_distances}
\alias{pbc_distances}
\title{Calculating distances between points subject to periodic boundary constraints.}
\usage{
pbc_distances(points, lims)
}
\description{
Calculates pairwise distances between points subject to periodic boundary constraints.
}

