% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.r
\name{plot.nspp}
\alias{plot.nspp}
\title{Plotting an estimated Palm intensity function.}
\usage{
\method{plot}{nspp}(x, emp = FALSE, breaks = NULL, ...)
}
\arguments{
\item{x}{A fitted model from \link{fit.ns}().}

\item{emp}{Logical, if \code{TRUE} then the empirical Palm
intensity is also plotted.}

\item{breaks}{The (approximate) number of points plotted.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Plots a fitted Palm intensity function from an object returned by
\link{fit.ns}().
}

