% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm.r
\docType{data}
\name{porpoise.data}
\alias{porpoise.data}
\title{Two-camera porpoise data.}
\format{
A list.
}
\usage{
porpoise.data
}
\description{
Synthetic data constructed from circle-back aerial survey data; see
Stevenson, Borchers, and Fewster (in press).
}
\references{
Stevenson, B. C., Borchers, D. L., and Fewster,
    R. M. (in press) Cluster capture-recapture to account for
    identification uncertainty on aerial surveys of animal
    populations. \emph{Biometrics}.
}
\keyword{datasets}
