% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim-ns.r
\name{sim.ns}
\alias{sim.ns}
\title{Simulating Neyman-Scott point process data}
\usage{
sim.ns(pars = NULL, lims = rbind(c(0, 1), c(0, 1)), rchild = rpois,
  non.siblings = NULL, plot = TRUE, ...)
}
\arguments{
\item{pars}{A named vector (or list?) of parameter values. Required
parameters are \code{D}, the density of parent points, and
\code{sigma}, for the square root of the main diagonal of the
variance-covariance matrix of the multivariate normal distribution
of children point locations around their parent.}

\item{lims}{A matrix with two colums, corresponding to the upper
and lower limits of each dimension, respectively.}

\item{rchild}{A function for the generation of random
variables from the distribution of the number of children for each
parent. Parameter values for this distribution are specified in
\code{...}. The first argument of this function must be named
\code{n}, and set the number of random values to generate. Built in
functions (e.g., \code{rpois}) are suitable.}

\item{non.siblings}{An argument specifying the type of sibling
information to return (not sure what this should look like yet).}

\item{plot}{Logical, if \code{TRUE}, simulated parent and children
point locations will be plotted (only for two dimensions).}

\item{...}{Further parameters for rchild.dist.}
}
\value{
A matrix containing simulated point locations.
}
\description{
Simulates point locations from a Neyman-Scott point process.
}

