% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-ns.r
\name{sim.twoplane}
\alias{sim.twoplane}
\title{Simulating two-dimensional two-plane whale survey data}
\usage{
sim.twoplane(D, sigma, S, l, w, b, t, C)
}
\arguments{
\item{D}{Animal density.}

\item{sigma}{Animal dispersion between the passing of planes.}

\item{S}{Mean duration of surface phase.}

\item{l}{The length of the transect flown (in km).}

\item{w}{The distance from the transect to which detection of
individuals on the surface is certain. This is equivalent to the
half-width of the detection zone.}

\item{b}{The distance from the transect to the edge of the area of
interest. Conceptually, the distance between the transect and the
furthest distance a whale could be on the passing on the first
plane and plausibly move into the detection zone by the passing of
the second plane.}

\item{t}{The lag between planes (in seconds).}

\item{C}{Mean dive-cycle duration (in seconds).}
}
\value{
A list containing observed animal locations and associated
    plane IDs.
}
\description{
Simulates observed whale locations and plane IDs from a two-plane
whale survey.
}

