% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.r
\name{summary.nspp}
\alias{summary.nspp}
\title{Summarising nspp model fits}
\usage{
\method{summary}{nspp}(object, ...)
}
\arguments{
\item{object}{A fitted model from \link{fit.ns}.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Provides a useful summary of the model fit.
}

